backend/pyproject.toml
[project]
name = "copywriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.5",
    "anthropic>=0.84.0",
    "chromadb>=0.5.23",
    "fastapi>=0.133.1",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "openai>=2.24.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "tavily-python>=0.7.21",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.23.0"
]
quality = [
    "ruff>=0.9.0",
    "bandit>=1.8.0",
    "pip-audit>=2.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.coverage.run]
omit = [
    "main.py",
    "test_media.py",
    "test_orq.py",
    "tests/*"
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]
ignore = ["S603", "S607", "S404"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[dependency-groups]
dev = [
    "pytest-vcr>=1.0.2",
]
